requirements.txt
astroid==2.4.2
colorama==0.4.3
isort==5.5.2
keyboard==0.13.5
lazy-object-proxy==1.4.3
mccabe==0.6.1
MouseInfo==0.1.3
numpy==1.19.2
opencv-python==4.4.0.42
Pillow==7.2.0
PyAutoGUI==0.9.50
PyGetWindow==0.0.8
pylint==2.6.0
PyMsgBox==1.0.8
pyperclip==1.8.0
pypiwin32==223
PyRect==0.1.4
PyScreeze==0.1.26
pytesseract==0.3.6
PyTweening==1.0.3
pywin32==228
six==1.15.0
toml==0.10.1
win10toast==0.9
wrapt==1.12.1
